# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/webhooks/stripe": {
      "post": {
        "description": "Process webhook events from Stripe payment gateway",
        "operationId": "WebhookController_handleStripeWebhook",
        "parameters": [
          {
            "name": "stripe-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully"
          }
        },
        "summary": "Handle Stripe webhook",
        "tags": [
          "Webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "GameTriggers API",
    "description": "The Gametriggers marketplace API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}